@@ -5,7 +5,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=3.0.0'
 
 dependencies:
   flutter:
@@ -32,6 +32,8 @@ dependencies:
   resize: ^1.0.0
   intl: ^0.18.1
   gap: ^3.0.1
+  http: ^1.2.1
+  oktoast: ^3.4.0
 
 dev_dependencies:
   flutter_test:
