@@ -1,14 +1,14 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:gap/gap.dart';
 import 'package:get/get.dart';
 import 'package:intl/intl.dart';
-import 'package:portfolio/core/constants/app_constants.dart';
 import 'package:portfolio/core/constants/color_constants.dart';
+import 'package:portfolio/core/constants/portfolio_data.dart';
+import 'package:portfolio/core/model/achievement_model.dart';
 import 'package:portfolio/core/model/data.dart';
 import 'package:portfolio/core/model/model.dart';
-import 'package:portfolio/core/helpers/app_helpers.dart';
-import 'package:portfolio/screens/desktop/data_item.dart';
 import 'package:resize/resize.dart';
 
 class AboutScreen extends StatelessWidget {
@@ -18,72 +18,135 @@ class AboutScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     final screenSize = MediaQuery.of(context).size;
 
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        Text(
-          'Leadership Experience',
-          style: TextStyle(
-            fontWeight: FontWeight.w600,
-            color: AppColor.textColor,
-            fontSize: 30.sp,
+    return Container(
+      width: screenSize.width,
+      padding: EdgeInsets.only(top: 70.sp),
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        mainAxisAlignment: MainAxisAlignment.start,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Text(
+            'About Me',
+            style: TextStyle(
+              fontSize: 30.sp,
+              fontWeight: FontWeight.w500,
+            ),
           ),
-        ),
-        Gap(20.sp),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisSize: MainAxisSize.max,
-          children: [
-            DataItem(
-              model: Data.por[0],
-              showSkill: true,
-              color: AppColor.primary,
+          Gap(30.sp),
+          Container(
+            width: screenSize.width / 1.6,
+            padding: EdgeInsets.symmetric(
+              horizontal: 25.sp,
+              vertical: 20.sp,
             ),
-            DataItem(
-              model: Data.por[1],
-              showSkill: true,
-              color: AppColor.primary,
+            decoration: BoxDecoration(
+              color: AppColor.box,
+              borderRadius: BorderRadius.circular(13.sp),
             ),
-          ],
-        ),
-        Gap(20.sp),
-        DataItem(
-          model: Data.por[2],
-          showSkill: true,
-          color: AppColor.primary,
-        ),
-        Gap(20.sp),
-        Text(
-          'Achievements',
-          style: TextStyle(
-            fontWeight: FontWeight.w600,
-            color: AppColor.textColor,
-            fontSize: 30.sp,
+            alignment: Alignment.center,
+            child: Html(
+              data:
+                  '<p style="font-size:${18.sp};text-align: center">${Data.about}</p>',
+            ),
           ),
-        ),
-        Gap(20.sp),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisSize: MainAxisSize.max,
-          children: [
-            AppHelper.achievementItem(
-              Data.techAchievements,
-              screenSize,
-              AppConstants.tech,
+          Gap(30.sp),
+          Row(
+            mainAxisSize: MainAxisSize.max,
+            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              SizedBox(
+                width: screenSize.width / 2.5,
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Text(
+                      'Technical achievements',
+                      style: TextStyle(
+                        fontSize: 22.sp,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    ),
+                    Gap(30.sp),
+                    ListView.builder(
+                      physics: const NeverScrollableScrollPhysics(),
+                      padding: EdgeInsets.zero,
+                      shrinkWrap: true,
+                      itemCount: PortfolioData.techAchievements.length,
+                      itemBuilder: (ctx, index) => achievements(
+                        PortfolioData.techAchievements[index],
+                        screenSize,
+                        index != 0 ? 30.sp : 0,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              SizedBox(
+                width: screenSize.width / 2.5,
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Text(
+                      'Cultural achievements',
+                      style: TextStyle(
+                        fontSize: 22.sp,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    ),
+                    Gap(30.sp),
+                    ListView.builder(
+                      physics: const NeverScrollableScrollPhysics(),
+                      padding: EdgeInsets.zero,
+                      shrinkWrap: true,
+                      itemCount: PortfolioData.culAchievements.length,
+                      itemBuilder: (ctx, index) => achievements(
+                        PortfolioData.culAchievements[index],
+                        screenSize,
+                        index != 0 ? 30.sp : 0,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+          Gap(30.sp),
+          Text(
+            'Hey there! Ready to be blown away? Check out my Instagram for some killer singing and guitar skills! 🎶🎸',
+            style: TextStyle(
+              fontWeight: FontWeight.w400,
+              color: AppColor.textColor,
+              fontSize: 18.sp,
             ),
-            AppHelper.achievementItem(
-              Data.culAchievements,
-              screenSize,
-              AppConstants.cul,
-            ),
-          ],
-        ),
-        Gap(20.sp),
-      ],
+          ),
+          Gap(30.sp),
+          // Gap(20.sp),
+          // Row(
+          //   mainAxisAlignment: MainAxisAlignment.spaceAround,
+          //   crossAxisAlignment: CrossAxisAlignment.center,
+          //   mainAxisSize: MainAxisSize.max,
+          //   children: [
+          //     AppHelper.achievementItem(
+          //       Data.techAchievements,
+          //       screenSize,
+          //       AppConstants.tech,
+          //     ),
+          //     AppHelper.achievementItem(
+          //       Data.culAchievements,
+          //       screenSize,
+          //       AppConstants.cul,
+          //     ),
+          //   ],
+          // ),
+          // Gap(20.sp),
+        ],
+      ),
     );
   }
 
@@ -181,6 +244,71 @@ class AboutScreen extends StatelessWidget {
           ),
         );
       },
+    );
+  }
+
+  Widget achievements(
+    AchievementModel achievement,
+    Size screenSize,
+    double upperMargin,
+  ) {
+    return Container(
+      width: screenSize.width / 2.5,
+      margin: EdgeInsets.only(top: upperMargin),
+      constraints: BoxConstraints(maxHeight: screenSize.height / 8),
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(13.sp),
+        color: AppColor.box,
+      ),
+      padding: EdgeInsets.symmetric(
+        horizontal: 20.sp,
+        vertical: 15.sp,
+      ),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.start,
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisSize: MainAxisSize.max,
+        children: [
+          Text(
+            achievement.position,
+            style: TextStyle(
+              fontSize: 22.sp,
+              fontWeight: FontWeight.w600,
+              color: AppColor.secondary,
+            ),
+          ),
+          Gap(30.sp),
+          Expanded(
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: [
+                Text(
+                  achievement.eventName,
+                  style: TextStyle(
+                    fontSize: 18.sp,
+                    fontWeight: FontWeight.w500,
+                    color: AppColor.secondary,
+                  ),
+                ),
+                Flexible(
+                  child: AutoSizeText(
+                    achievement.description,
+                    textAlign: TextAlign.center,
+                    style: TextStyle(
+                      fontSize: 16.sp,
+                      fontWeight: FontWeight.w400,
+                    ),
+                    minFontSize: 5,
+                    stepGranularity: 0.1,
+                  ),
+                ),
+              ],
+            ),
+          )
+        ],
+      ),
     );
   }
 }
