@@ -1,22 +1,53 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_html/flutter_html.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:gap/gap.dart';
-import 'package:intl/intl.dart';
+import 'package:get/get.dart';
 import 'package:portfolio/core/constants/app_constants.dart';
-import 'package:portfolio/core/constants/color_constants.dart';
-import 'package:portfolio/core/model/data.dart';
-import 'package:portfolio/core/model/model.dart';
-import 'package:portfolio/screens/desktop/data_item.dart';
+import 'package:portfolio/core/constants/portfolio_data.dart';
+import 'package:portfolio/core/helpers/app_helpers.dart';
 import 'package:portfolio/widgets/connect_button.dart';
+import 'package:portfolio/widgets/desktop/project_item.dart';
 import 'package:resize/resize.dart';
-import 'package:url_launcher/url_launcher_string.dart';
 
-class ProjectScreen extends StatelessWidget {
+class ProjectScreen extends StatefulWidget {
   const ProjectScreen({super.key});
+
+  @override
+  State<ProjectScreen> createState() => _ProjectScreenState();
+}
+
+class _ProjectScreenState extends State<ProjectScreen>
+    with SingleTickerProviderStateMixin {
+  late final AnimationController animationController;
+  double duration = 0;
+
+  @override
+  void initState() {
+    duration = PortfolioData.projects.length * 300;
+
+    animationController = AnimationController(
+      vsync: this,
+      duration: duration.milliseconds,
+    );
+
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    animationController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
     final screenSize = MediaQuery.of(context).size;
+
+    SchedulerBinding.instance.addPostFrameCallback((timeStamp) {
+      if (!animationController.isCompleted) {
+        animationController.forward();
+      }
+    });
 
     return Container(
       width: screenSize.width,
@@ -30,103 +61,32 @@ class ProjectScreen extends StatelessWidget {
           Text(
             'Projects',
             style: TextStyle(
-              fontWeight: FontWeight.w600,
-              color: AppColor.textColor,
+              fontWeight: FontWeight.w500,
               fontSize: 30.sp,
             ),
           ),
           Gap(30.sp),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisSize: MainAxisSize.max,
-            children: [
-              DataItem(
-                model: Data.projects[0],
-                showSkill: true,
-                color: AppColor.primary,
-              ),
-              DataItem(
-                model: Data.projects[1],
-                showSkill: true,
-                color: AppColor.primary,
-              ),
-            ],
-          ),
-          Gap(30.sp),
-          Text(
-            'Side Projects',
-            style: TextStyle(
-              fontWeight: FontWeight.w600,
-              color: AppColor.textColor,
-              fontSize: 30.sp,
+          Container(
+            width: screenSize.width,
+            padding: EdgeInsets.symmetric(horizontal: 30.sp),
+            child: Wrap(
+              runAlignment: WrapAlignment.start,
+              alignment: WrapAlignment.spaceEvenly,
+              spacing: 30.sp,
+              runSpacing: 30.sp,
+              children: buildChildren(),
             ),
           ),
           Gap(30.sp),
-          DataItem(
-            model: Data.projects[2],
-            showSkill: true,
-            color: AppColor.primary,
-          ),
-          Gap(30.sp),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisSize: MainAxisSize.max,
-            children: [
-              DataItem(
-                model: Data.projects[3],
-                showSkill: true,
-                color: AppColor.primary,
-              ),
-              DataItem(
-                model: Data.projects[4],
-                showSkill: true,
-                color: AppColor.primary,
-              ),
-            ],
-          ),
-          Gap(30.sp),
-          DataItem(
-            model: Data.projects[5],
-            showSkill: true,
-            color: AppColor.primary,
-          ),
-          Gap(30.sp),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisSize: MainAxisSize.max,
-            children: [
-              DataItem(
-                model: Data.projects[6],
-                showSkill: true,
-                color: AppColor.primary,
-              ),
-              DataItem(
-                model: Data.projects[7],
-                showSkill: true,
-                color: AppColor.primary,
-              ),
-            ],
-          ),
-          Gap(30.sp),
-          DataItem(
-            model: Data.projects[8],
-            showSkill: true,
-            color: AppColor.primary,
-          ),
-          Gap(30.sp),
           Text(
-            'More projects on Github',
+            'Unlock the treasure trove of my brilliance – behold all my projects on GitHub!',
             style: TextStyle(
-              fontWeight: FontWeight.w600,
-              color: AppColor.textColor,
-              fontSize: 26.sp,
+              fontWeight: FontWeight.w400,
+              fontSize: 18.sp,
             ),
           ),
           Gap(30.sp),
-          ConnectButton(
+          SocialButton(
             icon: AppConstants.githubAvatar,
             link: AppConstants.githubLink,
             size: Size(screenSize.width / 8, screenSize.width / 8),
@@ -137,83 +97,26 @@ class ProjectScreen extends StatelessWidget {
     );
   }
 
-  Widget projectItem(Model model, Size screenSize) {
-    bool hover = false;
-    return StatefulBuilder(builder: (_, setState) {
-      return MouseRegion(
-        opaque: false,
-        cursor: SystemMouseCursors.click,
-        child: InkWell(
-          splashColor: Colors.transparent,
-          onHover: (hovering) => setState(() {
-            hover = hovering;
-          }),
-          onTap: () => launchUrlString(model.link),
-          child: Container(
-            width: screenSize.width / 2.4,
-            padding: EdgeInsets.all(10.sp),
-            decoration: BoxDecoration(
-              color: AppColor.box,
-              borderRadius: BorderRadius.circular(10.sp),
-              boxShadow: hover
-                  ? [
-                      BoxShadow(
-                        color: AppColor.primary,
-                        spreadRadius: 1.sp,
-                        blurRadius: 30.sp,
-                      ),
-                    ]
-                  : [],
-            ),
-            alignment: Alignment.center,
-            child: Stack(
-              children: [
-                Column(
-                  mainAxisSize: MainAxisSize.min,
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Text(
-                      model.name,
-                      textAlign: TextAlign.center,
-                      style: TextStyle(
-                        fontWeight: FontWeight.w500,
-                        color: AppColor.textColor,
-                        fontSize: 25.sp,
-                      ),
-                    ),
-                    Gap(5.sp),
-                    Text(
-                      model.location,
-                      style: TextStyle(
-                        fontWeight: FontWeight.w400,
-                        color: AppColor.textColor.withOpacity(0.6),
-                        fontSize: 18.sp,
-                      ),
-                    ),
-                    Gap(5.sp),
-                    Text(
-                      '${DateFormat('MMMM yyyy').format(model.from)} - ${DateFormat('MMMM yyyy').format(model.to)}',
-                      style: TextStyle(
-                        fontWeight: FontWeight.w400,
-                        color: AppColor.textColor.withOpacity(0.6),
-                        fontSize: 15.sp,
-                        fontStyle: FontStyle.italic,
-                      ),
-                    ),
-                    Gap(5.sp),
-                    Html(
-                      data: '<p style="font-size:${20.sp};text-align: center">'
-                          '${model.description.join('<br><br>')}'
-                          '</p>',
-                    ),
-                  ],
-                ),
-              ],
-            ),
-          ),
+  List<Widget> buildChildren() {
+    final children = <Widget>[];
+    double interval = (duration / PortfolioData.projects.length)
+        .remap(0, duration, 0, 1)
+        .toDouble();
+    double sum = 0;
+
+    for (var index = 0; index <= PortfolioData.projects.length - 1; index++) {
+      children.add(
+        ProjectItem(
+          project: PortfolioData.projects[index],
+          begin: sum,
+          end: sum + interval,
+          animationController: animationController,
         ),
       );
-    });
+
+      sum += interval - (interval / 3);
+    }
+
+    return children;
   }
 }
